Revert webpack filer til å være js
